Feature/OT27-113 test endpoints  users

Approved-by: Franco Baisch
Approved-by: Gabriel Gomez
Approved-by: Emiliano Rivarola
Approved-by: Jonathan Oscar, Molina Gomez

+++ controllers/users.js
@@ -6,7 +6,11 @@ const { signToken, decodeToken } = require('../utils/jsonwebtoken');
 
 const infoUser = async (req, res) => {
   try {
-    const { id } = decodeToken(req, res);
+    const token = decodeToken(req );
+    if (!token)
+      return res.status(401).json({ error: 'token invalid or missing' });
+
+    const { id } = token;
     const user = await User.findOne({
       attributes: ['firstName', 'lastName', 'email', 'id', 'roleId'],
       where: {
@@ -14,47 +18,42 @@ const infoUser = async (req, res) => {
       }
     });
 
-    if (!user) {
-      res.status(404).json({ err: 'User not found' });
-    } else {
-      res.status(202).json(user);
-    }
+    if (!user) return res.status(404).json({ err: 'User not found' });
+
+    return res.status(200).json(user);
   } catch (e) {
     res.status(500).json({ error: e });
   }
 };
 
 const registerUser = async (req, res) => {
-
   const { email, firstName, image = null, lastName, roleId } = req.body;
 
   const userExists = await User.findOne({ where: { email: req.body.email } });
 
   try {
     if (!userExists) {
-
       const salt = bcrypt.genSaltSync(10);
       const hash = bcrypt.hashSync(req.body.password, salt);
-      
-        const newUser = await User.create({
-          firstName,
-          image,
-          lastName,
-          email,
-          password: hash,
-          roleId
-        });
 
-        const { password, ...dataForToken } = newUser.dataValues;
-        signToken(dataForToken, res);
+      const newUser = await User.create({
+        firstName,
+        image,
+        lastName,
+        email,
+        password: hash,
+        roleId
+      });
 
+      const { password, ...dataForToken } = newUser.dataValues;
+      signToken(dataForToken, res);
     } else {
       res.status(409).json({
         message: 'User already registered'
       });
     }
   } catch (err) {
-    console.log( err )
+    console.log(err);
     res.status(500).json({
       message: err
     });
@@ -62,13 +61,15 @@ const registerUser = async (req, res) => {
 };
 
 const deleteUser = async (req, res) => {
-  const id = req.params.id;
-
+  const token = decodeToken(req);
+  if (!token)
+    return res.status(401).json({ error: 'token invalid or missing' });
+  const { id } = token;
   try {
     await User.destroy({ where: { id } });
     return res.status(200).json({ message: 'User deleted successfuly' });
   } catch (error) {
-    res.status(400).send(error.message);
+    res.status(500).json(error);
   }
 };
 
@@ -78,16 +79,16 @@ const getUsers = async (req, res) => {
       include: [
         {
           model: Role,
-          as: 'role'
+          as: 'role',
+          attributes: ['name']
         }
       ],
-      attributes: ['id', 'firstName', 'lastName', 'email', 'image']
+      attributes: ['firstName', 'lastName', 'roleId']
     });
 
-    if (userList.length === 0) throw new Error('The resources do not exist');
     res.status(200).json(userList);
   } catch (error) {
-    res.status(400).json({ status: 400, error: error.message });
+    res.status(500).json({ error: error.message });
   }
 };
